+++ indra/tools/reading/pmid_reading/__init__.py
@@ -8,37 +8,15 @@ import pickle
 import logging
 import argparse
 from datetime import datetime
-from platform import system
-import read_sparser
-import read_reach
+from .util import get_proc_num
+from . import read_sparser
+from . import read_reach
 
 READER_DICT = {'reach': read_reach.run_reach,
                'sparser': read_sparser.run_sparser}
 
 
 logger = logging.getLogger('runreader')
-
-
-def get_proc_num():
-    if system() == 'Linux':
-        with open('/proc/cpuinfo', 'r') as f:
-            ret = len([
-                line for line in f.readlines() if line.startswith('processor')
-                ])
-    else:
-        ret = None
-    return ret
-
-
-def get_mem_total():
-    if system() == 'Linux':
-        with open('/proc/meminfo', 'r') as f:
-            lines = f.readlines()
-        tot_entry = [line for line in lines if line.startswith('MemTotal')][0]
-        ret = int(tot_entry.split(':')[1].replace('kB', '').strip())/10**6
-    else:
-        ret = None
-    return ret
 
 
 parser = argparse.ArgumentParser(
